validasi insert and deletion

+++ menu.cpp
@@ -30,6 +30,7 @@ void menu_utama(){
 	BackMenu:
 	menu_list();
 	printTree(root_node);
+	cout << endl;
 //	fflush(stdin); 
 			  heapSortAscending(arr, count);
 			  setcolor(2);
@@ -47,19 +48,28 @@ void menu_utama(){
 //	fflush(stdin);
  	switch(inp){
  		case 1:
- 			printf("%d",count);
+// 			printf("%d",count);
  			while(data != 100){
 			cout << "Data : ";
 			cin >> data ;
-			arr[count] = data;			
-			count++;
-			if(root_node == NULL){
-				root_node = insertNode(root_node, data);
+			if(checkNode(root_node, data)){
+				setcolor(4);
+				cout << "Data sudah ada di dalam tree" << endl;
+				setcolor(7);
+				system("pause");
+				system("cls");
+				goto BackMenu;
 			}else{
-				insertNode(root_node, data);
+				arr[count] = data;			
+				count++;
+				if(root_node == NULL){
+					root_node = insertNode(root_node, data);
+				}else{
+					insertNode(root_node, data);
+				}
+				system("CLS");
+				goto BackMenu;
 			}
-			system("CLS");
-			goto BackMenu;
 		}
 			break;
 		case 2 :
@@ -70,19 +80,28 @@ void menu_utama(){
 			data = NULL;
 			menu_utama();
 			break;
-		case 3:
-//			fflush(stdin);
+		case 3:	
 			cout << "Data : ";
 			cin >> data;
-			root_node = deleteNode(root_node, data);
-			DeleteArray(arr,data,count);
-//			system("cls");
-			system("pause");
-			setcolor(2);		
-			cout << "DELETE NODE " << endl;
-			setcolor(7);	
-			goto BackMenu;
-			break;
+			if(checkNode(root_node, data)){
+				root_node = deleteNode(root_node, data);
+				DeleteArray(arr,data,count);
+	//			system("cls");
+				system("pause");
+				setcolor(2);		
+				cout << "DELETE NODE " << endl;
+				setcolor(7);	
+				goto BackMenu;
+				break;
+			}else{
+				setcolor(4);
+				cout << "Data tidak ada di dalam tree" << endl;
+				setcolor(7);
+				system("pause");
+				system("cls");
+				goto BackMenu;
+			}
+
 		case 4:
 			  heapSortAscending(arr, count);
 			  setcolor(2);
